Has changed keys handling
