+++ src/main/java/sayner/sandbox/neuralG/textures/Texture.java
@@ -1,4 +1,4 @@
-package sayner.sandbox.NeuralG.Textures;
+package sayner.sandbox.neuralG.textures;
 
 import static org.lwjgl.opengl.GL11.*;
